feature/AB#82 Fixed DataTypes.js Moved DataTypes to EntityDialog.jsx
