--- webclient/app/src/app/config/DataTypes.js
@@ -1,42 +1,0 @@
-const DataTypes = [
-    {
-        id: 1,
-        name: "String"
-    },
-    {
-        id: 2,
-        name: "Integer",
-        allowMin: true,
-        allowMax: true
-    },
-    {
-        id: 3,
-        name: "Double",
-        allowMin: true,
-        allowMax: true
-    },
-    {
-        id: 4,
-        name: "Date"
-    },
-    {
-        id: 5,
-        name: "Time"
-    },
-    {
-        id: 6,
-        name: "Timestamp"
-    },
-    {
-        id: 7,
-        name: "BigDecimal",
-        allowMin: true,
-        allowMax: true
-    },
-    {
-        id: 8,
-        name: "Enum"
-    }
-];
-
-export default DataTypes;
