+++ webclient/app/src/app/features/apps/entities/entityDialog/DataTypes.js
@@ -1,0 +1,45 @@
+const DataTypes = [
+    {
+        id: 1,
+        name: "String",
+        allowMin: true,
+        allowMax: true,
+        usesLengthLimit: true
+    },
+    {
+        id: 2,
+        name: "Integer",
+        allowMin: true,
+        allowMax: true
+    },
+    {
+        id: 6,
+        name: "Long",
+        allowMin: true,
+        allowMax: true
+    },
+    {
+        id: 7,
+        name: "BigDecimal",
+        allowMin: true,
+        allowMax: true
+    },
+    {
+        id: 4,
+        name: "Float",
+        allowMin: true,
+        allowMax: true
+    },
+    {
+        id: 3,
+        name: "Double",
+        allowMin: true,
+        allowMax: true
+    },
+    {
+        id: 5,
+        name: "Boolean"
+    }
+];
+
+export default DataTypes;
